[refactor] move home_screen widget after rename (#47)
